backend/requirements.txt
asgiref==3.11.1
Django==6.0.2
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
packaging==26.0
pillow==12.1.0
psycopg2-binary==2.9.11
PyJWT==2.11.0
python-dotenv==1.2.1
setuptools==82.0.0
sqlparse==0.5.5
tzdata==2025.3
wheel==0.46.3
